fix(openclaw): cast session seq to unsigned for %u format
uint32_t is `long unsigned int` on xtensa (and unsigned int on host),
so passing it directly to %u triggers -Werror=format under the Xtensa
toolchain. Cast at the call site.

+++ main/application.cc
@@ -1194,7 +1194,8 @@ static std::string CurrentOpenclawUserId() {
         return CONFIG_OPENCLAW_USER;
     }
     char buf[64];
-    snprintf(buf, sizeof(buf), "%s-s%u", CONFIG_OPENCLAW_USER, seq);
+    snprintf(buf, sizeof(buf), "%s-s%u",
+             CONFIG_OPENCLAW_USER, static_cast<unsigned>(seq));
     return buf;
 }
